Remove special case treatment of zero width space
This ancient special case hack was intended to account for
differences between simple path and complex path.
isZeroWidthSpaceGlyph() was only used in HarfBuzzShaper, and this
function has a non-null-glyph check in it, so overwriting the
zero width space glyph with 0 does not seem to be a meaningful
operation anymore.

Turns out the line-end whitespace changes on Mac are actually
progressions, and those are shown on Windows, too, as an
indication that the selection continues on the next line. The
previous whitespace size modifications suppressed showing this.

BUG=566031
R=eae,bungeman2
TEST=fast/text/wide-zero-width-space.html

Review URL: https://codereview.chromium.org/1509973002

Cr-Original-Commit-Position: refs/heads/master@{#364115}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 847018eb85c219366f07e0c2b63fddc046c09ad9