Changing the equation of computing sigma from blur radius
This is implemented in skBlurRadiusToSigma, the equation
is: sigma = factor*radius + 0.5. The number 0.5 does not
make sense and this CL removes it. But that means this CL
will change the blur kernel, which results in NeedsRebaseline
for ~100 layout tests.

BUG=503626

Review URL: https://codereview.chromium.org/1502283002

Cr-Original-Commit-Position: refs/heads/master@{#364124}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8370a4fbe511ecc3a638f1448843015ae92160ca