Fix CrossesExtensionExtents to handle hosted apps with --isolate-extensions.
Currently, when determining whether a process swap is necessary for a
new window that was just window.open'ed, CrossesExtensionExtents
checks for --site-per-process and --isolate-extensions flags to
determine whether to look at an opener or the opener's top frame for
the source url.  This logic is busted with --isolate-extensions
and hosted apps, since --isolate-extensions doesn't isolate hosted
apps.  This caused AppApiTest.OpenAppFromIframe to fail for
--isolate-extensions.

This CL changes CrossesExtensionExtents to instead check the opener's
local frame root for this scenario, which should work for all current
and future Site Isolation modes.  To facilitate this, the local frame
root is now exposed from WebLocalFrame.

BUG=532666
TEST=AppApiTest.OpenAppFromIframe for --isolate-extensions.

Review URL: https://codereview.chromium.org/1511003002

Cr-Original-Commit-Position: refs/heads/master@{#364147}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0ab0a6e4adebd1920cfd322c1bb90d16f1738fea