Tweaking WebPageSerializerImpl to emit a BOM for UTF16/32.
Previously save-page-as would truncate saved html files if they
would include a '\0' byte (i.e. if saving a html document that was
encoded as utf16).  The truncation doesn't happen anymore [1], but the
saved html files still don't round-trip properly, because the browser is
not able to detect the encoding of the files (<meta> character cannot be
sniffed in encodings that are not ascii compatible [2] + it is obviously
impossible for local files to declare their encoding via http headers).

This CL makes changes to always emits a BOM when saving in UTF16
or UTF32 encoding.  This in-turn causes round-tripping tests in
save_page_browsertest.cc to pass for tests using these encodings.

Note that this CL applies only to save as "complete html" and has no
impact to save as "html only" or save as "mhtml".

[1] https://chromium.googlesource.com/chromium/src/+/17e4a38aef61b832589933c942426a9818e008ec%5E%21/content/renderer/render_frame_impl.cc

[2] http://www.w3.org/TR/html5/document-metadata#charset:
    If an HTML document contains a meta element with a charset attribute
    or a meta element with an http-equiv attribute in the encoding
    declaration state, then the character encoding used must be an
    ASCII-compatible character encoding.

BUG=541699

Review URL: https://codereview.chromium.org/1407663004

Cr-Original-Commit-Position: refs/heads/master@{#364394}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 36f0d157b6d2c6fae627a3aba955ad4e0fe07d71