Remove the scheduler from the PPAPI thread
The scheduler is designed to be the only thing talking to the underlying
base messageloop.  We arguably don't need it on the PPAPI thread so this
patch removes it and fixes the origional bug that prompted its use in a
different way.

BUG=558825, 534972

Review URL: https://codereview.chromium.org/1511893003

Cr-Original-Commit-Position: refs/heads/master@{#364688}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f11baf32d47858ede5bb8a5546388d1b501a10e3