The replaced boolean on LayoutObject is misnamed
'replaced' in CSS has a meaning and is mapped to inheriting
from LayoutReplaced in our code (see
https://codereview.chromium.org/1329923002). However we
also had a boolean named 'replaced' on LayoutObject, which
was super confusing.

To clear the confusion, renamed the 'replaced' boolean to
'atomic inline-level box' as this is what it maps to.

Review URL: https://codereview.chromium.org/1425743004

Cr-Original-Commit-Position: refs/heads/master@{#364715}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 893d17e2c6d6386e4b714962764ea2f9efa30cf2