Add setBufferingStrategy() to WebMediaPlayer.
This allows Blink to control when aggressive buffering triggers.

The initial implementation requires a user gesture to trigger aggressive buffering on pause. This is more restricted than the current Chromium behavior to trigger on any pause, in particular for pauses that occur as a result of removing an element from the DOM.

BUG=538523

Review URL: https://codereview.chromium.org/1489233004

Cr-Original-Commit-Position: refs/heads/master@{#364890}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c6ab163a49aa21bb2732e7a790c005b7a40fef89