Remove unused Blink accessibility code.
Most of this is code from WebKit we never used, or stuff we recently
replaced like the name calculation code. If any of it seems potentially
useful it will still be easy to recover from git, but in the meantime
removing it will be better for readability.

BUG=none

Review URL: https://codereview.chromium.org/1526773003

Cr-Original-Commit-Position: refs/heads/master@{#365186}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4c3dcc0ff1ffc3d21b0ece662c447cd033242e60