Remove check for invisible subtrees in LayoutGeometryMap.
It was added in https://codereview.chromium.org/27104002 to fix a problem wherein
some layer positions were not calculated in PaintLayer because of invisibility.

The code has changed a lot in PaintLayer and in compositing (for the better) since
then. I don't think this is necessary, and it's causing a lifecycle violation
assert elsewhere.

BUG=566936

Review URL: https://codereview.chromium.org/1514243004

Cr-Original-Commit-Position: refs/heads/master@{#365303}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 19a84549ad0bec1308b768f96b2345d4431a8a30