Added eventSender chorded buttons state.
We currently have two mouse-button-related issues in eventSender:
- EventSender currently remembers mouse location and button-press info
as expected, but doesn't remember more than the last button press. This
makes it tricky to test chorded button presses because all currently
depresssed buttons needs to be passed as modifiers.
- When a mouse button is pressed, both "button" and "buttons" fields of
the fired MouseEvent should automatically reflect the pressed button.
Currently, only "button" field carries the info unless eventSender gets
an explicit modifier for the pressed button. In a few layout tests, we
had to pass modifiers as follows to make the test pass only because
the tested code relies on "buttons" not "button":
    eventSender.mouseDown(0, ["leftButton"])
This is counter-intuitive and error-prone.

This CL adds a "buttons" state in eventSender to fix the problems.

BUG=563676

Review URL: https://codereview.chromium.org/1526583004

Cr-Original-Commit-Position: refs/heads/master@{#365343}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 21aed4eac48e492044aa079b4804ce90d894ccf7