Make window.focus work for OOPIF.
This CL adds the plumbing necessary for window.focus to work with
out-of-process iframes.  There are two things that need to happen when
calling window.focus() on a remote frame.  First, the target frame
needs to be set as focused in all renderers in its FrameTree.  Second,
if the new focused frame is a top-level frame in a different tab, that
tab should be activated and brought to front.  The latter happens via
page->chromeClient().focus() and is only allowed in some situations
(when window interaction is allowed and for popups whose opener points
to the frame that's executing window.focus).

Specific changes:

1) focus() is made non-virtual and moved from LocalDOMWindow to
DOMWindow, where it handles calls on both local and remote DOMWindows.

2) FrameClient::frameFocused now handles both local and remote focused
frame changes.  The plumbing for the local case is moved from
ChromeClientImpl.  The plumbing for the remote case is added so that
calling window.focus() on a remote frame sends
FrameHostMsg_FrameFocused from that frame's RenderFrameProxy to the
browser process.

3) FrameTree::SetFocusedFrame now takes a source SiteInstance
(representing the renderer process which initiated the focus change),
and for cross-process window.focus situations, it sends a
FrameMsg_SetFocusedFrame to the appropriate RenderFrame telling it to
focus.

BUG=341918, 339659

Review URL: https://codereview.chromium.org/1414853006

Cr-Original-Commit-Position: refs/heads/master@{#365623}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5357efb7505b2b56226c1d7c8fcaafab5fcdc221