Stylesheet rule cssText is all lowercase in quirks mode
We currently set the selector value to lowercase in quirks mode. FF
does not do this. This patch splits up the value stored in RareData
such that the one used for matching is maintained in lower case in
quirks mode. The serialized value is retained as it was provided.

BUG=74750

Review URL: https://codereview.chromium.org/1520423004

Cr-Original-Commit-Position: refs/heads/master@{#366008}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1860234300a5b0c7ba9c25b52bec62096cb475f5