XHR for Suborigin should only preflight if non-simple headers.
The status quo was sending a preflight for every XHR request from a
suborigin, regardless of the type of headers. This is overly aggressive
and should only be done for non-simple headers.

This modifies XHR to do that and adds a variety of tests to verify.

BUG=336894
R=jochen@chromium.org

Review URL: https://codereview.chromium.org/1536903002

Cr-Original-Commit-Position: refs/heads/master@{#366272}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 76d068c940fa13c23a4a4f3816a79358cf8a11cd