Ensure zoom level is not updated if main frame is WebRemoteFrame.
When swapped out RenderFrame is not used, the RenderView/WebView does not
have a main WebLocalFrame. This CL ensures the WebView code is not called
in such cases until it is updated to properly handle out-of-process iframes.

BUG=571603

Review URL: https://codereview.chromium.org/1546603003

Cr-Original-Commit-Position: refs/heads/master@{#366698}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eaa6adc9f2b8463b7f5859e3b9faffecdd5225f1