Fix error in determining the target position for the smooth scroll animation curve.
This is for the user scroll events that have a
mouse wheel handler and get handled on the main thread.

This bug was introduced in crrev.com/1534813004, where when we
get consecutive scroll events, the first one overwritten if it
is waiting to be sent to the the compositor.

BUG=552556

Review URL: https://codereview.chromium.org/1547003002

Cr-Original-Commit-Position: refs/heads/master@{#366799}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d0a670ec608697b975b7ca8e5bf9ef976afd4862