Remove ancestor walking code in SelectorFilter.
The code in setupParentStack would fill the selector filter with ancestor data
when a pushParent call was made with an empty parent stack and the element was
not the documentElement. Post lazy attach we always start style recalc from
the documentElement, and never pushParent outside recalcStyle, so this is just
dead code.

Lets remove the code and simplify the checks.

Review URL: https://codereview.chromium.org/1556183002

Cr-Original-Commit-Position: refs/heads/master@{#367484}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f3864bc03cf260ac3e68b0de8b97a1435d0be028