Crash in LayoutObject::isRooted during paint invalidation
The issue is that LayoutPart crosses frame boundaries, thus we
could have a frame in a different state than the main frame.
Ideally we would have better tightness between frames' states
instead of requiring this code.

As the test is flaky and I couldn't reproduce the issue locally,
it was better not to include it.

BUG=566247

Review URL: https://codereview.chromium.org/1555303002

Cr-Original-Commit-Position: refs/heads/master@{#367584}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 29b274ba041d2d98e4b34a0ff5f054b6c7bbdf54