Update optimized LayoutBlockFlowLine width computation criteria
Change the criteria triggering the optimized path in LayoutBlockFlowLine
to situations where word based shaping may be used. This is more correct
and consistent with how text layout and metrics works elsewhere. In some
situations this may result in a performance regression as correctness is
traded for speed. In the common case though this should result in faster
rendering as a character-by-character scan of the string can be avoided.

BUG=545147
R=pdr@chromium.org

Review URL: https://codereview.chromium.org/1565453002

Cr-Original-Commit-Position: refs/heads/master@{#367861}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b5ac0ffbecb471f903cc46da12be4d725cd0414d