Layered table row/cell leaves collapsed border area transparent
Collapsed borders are painted on the table's layer. If a row or a cell
has layer and the table collapses borders, the row or cell leaves the
collapsed border area on the layer transparent, letting collapsed borders
painted on the table's layer be seen through. In the case, the layer for
the row/cell should not be treated as opaque layers.

This change affects whether the layers are treated as opaque when the
layers are composited.

BUG=562333
TEST=LayoutTableRowTest.BackgroundIsKnownToBeOpaqueWithLayerAndCollapsedBorder
TEST=LayoutTableCellTest.BackgroundIsKnownToBeOpaqueWithLayerAndCollapsedBorder

Review URL: https://codereview.chromium.org/1547563002

Cr-Original-Commit-Position: refs/heads/master@{#367995}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bbe3833cab208c47f91dfba5decb356bba8f3dfb