Crash in LayoutObject::isRooted
The issue is that when the paint invalidation code recurses into
children frames in FrameView, there is no guarantees that the
children frames are in a state to generate paint invalidations
(e.g. missing a LayoutView as they are stopping).
This crash is responsible for 0.5% of the renderer crashes
but unfortunately I couldn't get a reproduction.

BUG=575146

Review URL: https://codereview.chromium.org/1570483002

Cr-Original-Commit-Position: refs/heads/master@{#368093}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f0b40bb2ba3419df0f763452f19841d4df2e3df1