Clear integrity failure bit in PendingScript on release
This fixes a bug where a script loaded after a previous script failed
due to an integrity failure would also fail to load. This is because the
PendingScript object was reused for the new script resource, and we
failed to clear out the m_integrityFailure bit. This simply sets that to
false when the old script resource is released.

BUG=570340

Review URL: https://codereview.chromium.org/1569823004

Cr-Original-Commit-Position: refs/heads/master@{#368379}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2a03de66caa66854a74e3e2fd006abbbff93ffb5