Force all experiment enabled checks to use ExperimentalFeatures
The original motivation for this CL was to add an overload
to the isApiEnabled() method, without an error message.
Like ExecutionContext.isSecureContext(), this allows
callers to check for enabled experiments, but not pass
an error message parameter, which would be ignored.

In adding the overload, it was determined that all callers should
be using the ExperimentalFeatures generated class
(with the exception of unit tests). As a result, the
Experiments class now just provides a private
implementation of the enabled check. Callers use the
ExperimentalFeatures class for the integrated check of
runtime-enabled + appropriate API key.

The motivation for this change was this CL to add the
[ExperimentEnabled] IDL attribute:
https://codereview.chromium.org/1531443003

BUG=551609

Review URL: https://codereview.chromium.org/1541983003

Cr-Original-Commit-Position: refs/heads/master@{#368471}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 707a221a0a22c5a376dcb7979367b4ee1c314f77